+++ tests/app/rdev/applications/bootstrappers/io/BootstrapperIOTest.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Copyright (C) 2015 David Young
+ *
+ * Tests the bootstrapper reader/writer
+ */
+namespace RDev\Applications\Bootstrappers\IO;
+
+class BootstrapperIOTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * Simply a placeholder
+     */
+    public function testTodo()
+    {
+        $this->assertTrue(true);
+    }
+}