+++ tests/app/rdev/applications/bootstrappers/mocks/LazyBootstrapper.php
@@ -2,14 +2,14 @@
 /**
  * Copyright (C) 2015 David Young
  *
- * Defines a bootstrapper that manipulates the environment
+ * Mocks a lazy bootstrapper
  */
 namespace RDev\Tests\Applications\Bootstrappers\Mocks;
 use RDev\Applications\Bootstrappers\Bootstrapper as BaseBootstrapper;
 use RDev\Applications\Bootstrappers\ILazyBootstrapper;
 use RDev\IoC\IContainer;
 
-class EnvironmentBootstrapper extends BaseBootstrapper implements ILazyBootstrapper
+class LazyBootstrapper extends BaseBootstrapper implements ILazyBootstrapper
 {
     /**
      * {@inheritdoc}
@@ -25,21 +25,5 @@ class EnvironmentBootstrapper extends BaseBootstrapper implements ILazyBootstrap
     public function registerBindings(IContainer $container)
     {
         $container->bind(LazyFooInterface::class, LazyConcreteFoo::class);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function run()
-    {
-        $this->environment->setName("running");
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function shutdown()
-    {
-        $this->environment->setName("shutting down");
     }
 }