+++ tests/app/rdev/applications/bootstrappers/mocks/NonBootstrapper.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Copyright (C) 2015 David Young
+ *
+ * Defines a class that is not a bootstrapper
+ */
+namespace RDev\Tests\Applications\Bootstrappers\Mocks;
+
+class NonBootstrapper
+{
+    // Don't do anything
+}